Add strip for dark actionbar+search bar.Switch refresh+search order
[Add]: Added orange strip 9-part drawable
[Mod]: Added attribute to hold the actionbar strip for the search bar
[Mod]: styles - 
-Set dark theme action bar to have orange strip
- Added drawable strip attribute to the dark and light themes. Light
theme set it to @null (transparent) while Dark theme set it to orange
strip
[Mod]: Added strip as background search bar dialog background

[Mod]: Switched order of refresh & search. This is for future
compatibility as it is likely planned that the "Career Item" fragment
will be capable of search (thus needs to be aligned right).
Additionally, it makes refreshing a more intentional action while search
may be pressed if the setting drop down is hit.