Fixed sroll+misfire issue + added a potential flickering fix
This commit addresses issue #11

[Fix]: CareerItemFragment
-The scroll issue was a two-fold issue. First cause by the fact that the
view height would be incomplete until the WebView loads AND as such the
scroll view ignores its height. To remedy it, the scroll is set using
the sum of scroll + webview heights, and is set on webview load
- The misfire issue is simply fixed by checking the height at load and,
if 0, it forces a reload. Reloads typically never exceed 2, but 5 is set
as a safety.

[Mod][Fix]: Potential fix for flicking on the WebView. This occurs when
2D Hardware acceleration is forced on some devices. To fix this no
transparency (the reason it was turned of) is used and the background
are set via careerstack_background_colour. This is not yet tested.