Implemented the tag search functionality Activity + CareerItemFrag
[Add][Mod][Fix]: CareerItemFragment -
- Added the class OnFragmentInteractionListener for sending tag requests
back to the activity
- Fixed a potential bug with loadJobDescription() and buildAndLoadTags()
in which they throw FragmentNotAttached. Basically just a pre-check in
both methods + some extra try-catching (just in case).
- Added the toggle button for the tags. Tags are shown or not based on
the checkbox button, sets a preference. Said button is made visible by a
resolution based bool
- Added delay to the job description loading when tags are being shown

[Mod]: Implemented OnFragmentInteractionListener in the activity; it
will launch a new search and set the search bar to use the tag as a
keyword

This should resolve issue #5