Fixed urgent bug regarding Search Frag's external searches
[Fix]: When SearchResultsFragment had search issued to it with a
destroyed view (e.g. after rotation), it would fail. This is because the
loading views would be invalid can cause (caught) null exceptions. The
solution is multi-faceted: 
1. Check that views are non-null (solves exceptions)
2. Use the state values for: prevQuery and loading bools only if said
variables are not already set (such as a current search)

The second part is important for displaying the correct loading
animations + correct search summaries

[Mod]: Updated version code/name for release (1.0.0)
    
Version 1.0.0:
-Added search tags
-Added fade animations to search + job descriptions
-Added background fade to the search bar dialog
-Added progress count to the search loading
-Fixed job descriptions scroll, misfirings, & flickering bugs